added support for alternate output file in tap
added two api functions which allow a file stream
to be specified in order to override the use of
stdout for normal output and stderr for error output.

the c ternary expression is used whenever the TAP_OUTPUT
and TAP_ERR_OUTPUT macro is evaluated in order to
avoid requiring an init function to be called which
would setup the pointers to the output streams.